split multi reinforcement commanders into modules
